pyproject.toml
[project]
name = "trade-agent"
version = "0.1.0"
description = "AI-powered trading assistant that analyzes portfolios, options positions, and market data from TradeView"
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.136.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
]

deployment = [
    "absl-py>=2.2.1",
]

[tool.ruff.lint]
ignore = ["C901"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["trade_agent"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["trade_agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
